Move Thread to rely less on shared examples
