+++ core/thread/exit_spec.rb
@@ -1,6 +1,11 @@
 require_relative '../../spec_helper'
 require_relative 'fixtures/classes'
-require_relative 'shared/exit'
+
+describe "Thread#exit" do
+  it "is an alias of Thread#kill" do
+    Thread.instance_method(:exit).should == Thread.instance_method(:kill)
+  end
+end
 
 describe "Thread#exit!" do
   it "needs to be reviewed for spec completeness"
