+++ core/thread/fork_spec.rb
@@ -1,9 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/start'
 
 describe "Thread.fork" do
-  describe "Thread.start" do
-    it_behaves_like :thread_start, :fork
+  it "is an alias of Thread.start" do
+    Thread.method(:fork).should == Thread.method(:start)
   end
 end
